geofence PointOfInterests
The user can now set discord-channel ids, geofences in the config file, they are defined as list.

If a search command is executed in a channel, the SearchEngine will filter out PointOfInterest according to the geofence which is assigned to it.

If a channel if not present, the search engine will not filter